// client/src/Interfaces/TeachersIn.ts
export interface ListBseUserIn {
    id: number
    name: string
    lastName: string
    email: string
    image: string
    role: string
}

export interface TeacherPrIn {
    id: number | null
    name: string | null
    lastName: string | null
    email: string | null
    image: string | null
    role: string | null
}